\name{boosttest-package}
\alias{boosttest-package}
\alias{boosttest}
\docType{package}
\title{
\packageTitle{boosttest}
}
\description{
\packageDescription{boosttest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{boosttest}
\packageIndices{boosttest}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{boosttest}

Maintainer: \packageMaintainer{boosttest}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
