% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cleanC.R
\name{cleanC}
\alias{clean}
\alias{clean.character}
\alias{cleanC}
\title{simple cleaning of text before processing}
\usage{
cleanC(x)

clean(x)

\method{clean}{character}(x)
}
\arguments{
\item{x}{character vector to be cleaned}
}
\value{
A character vector equal in length to the original texts, after cleaning.
}
\description{
simple cleaning of text before processing
}
\details{
minimal test \code{clean} that only removes digits and punctuation
for testing.
}
\examples{
cleanC("This is 1 sentence with 2.0 numbers in it, and one comma.")
cleanC("This is a $1,500,000 budget and $20bn cash and a $5 cigar.")

# for a vector of texts
cleanC(c("This is 1 sentence with 2.0 numbers in it, and one comma.",
        "$1.2 billion was spent on text analysis in 2014."))

\dontrun{# on a single long text
mobydick <- texts(corpus(textfile("~/Dropbox/QUANTESS/corpora/project_gutenberg/pg2701.txt")))
system.time(tmp <- cleanC(mobydick)) # .218 seconds
system.time(tmp <- clean(mobydick))  # .776 seconds}
clean("This is 1 sentence with 2.0 numbers in it, and one comma.")
clean("This is a $1,500,000 budget and $20bn cash and a $5 cigar.")

# for a vector of texts
clean(c("This is 1 sentence with 2.0 numbers in it, and one comma.",
        "$1.2 billion was spent on text analysis in 2014."))
}

